requirements.txt
streamlit==1.41.1
bs4==0.0.2
langchain==0.3.13
langchain-community==0.3.13
langchain-google-genai==2.0.8
python-dotenv==1.0.1